# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chama by Aventum - Digital Rotational Savings Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ORGANIZATIONS
...
  /payouts:
    get:
      operationId: listPayouts
      tags: [payouts]
      summary: List payouts for current user
      parameters:
        - in: query
          name: groupId
          schema:
            type: integer
      responses:
        "200":
          description: List of payouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payout"
  /payouts/all:
    get:
      operationId: listAllPayouts
      tags: [payouts]
      summary: List all payouts (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All payouts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutListResponse"
  /payouts/{payoutId}/complete:
    post:
      operationId: completePayout
      tags: [payouts]
      summary: Mark a payout as completed (admin)
      parameters:
        - in: path
          name: payoutId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payout completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [member, group_admin, org_admin, super_admin, ceo, cto_admin, it_support, finance, marketing, relationship_manager]
        organizationId:
          type: integer
          nullable: true
        departmentId:
          type: integer
          nullable: true
        isFinanceAdmin:
          type: boolean
          nullable: true
        twoFactorEnabled:
          type: boolean
        phoneNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        emailMarketing:
          type: boolean
          nullable: true
        notificationEmail:
          type: boolean
        notificationSms:
          type: boolean
        avatar:
          type: string
          nullable: true

    # ORGANIZATION
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        adminId:
          type: integer
        organizationId:
          type: integer
          nullable: true
        contributionAmount:
          type: number
        currency:
          type: string
          default: USD
        schedule:
          type: string
          enum: [weekly, bi-weekly, monthly]
        maxMembers:
          type: integer
        currentCycle:
          type: integer
        currentRotationIndex:
          type: integer
        status:
          type: string
          enum: [active, paused, completed]
        totalMembers:
          type: integer
        paidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminId
        - contributionAmount
        - schedule
        - maxMembers
        - currentCycle
        - currentRotationIndex
        - status
        - totalMembers
        - paidCount
        - createdAt
...
    Payout:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        cycleId:
          type: integer
        recipientId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, failed]
        paidAt:
          type: string
          format: date-time
          nullable: true
        recipient:
          $ref: "#/components/schemas/User"
          nullable: true
        group:
          $ref: "#/components/schemas/Group"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - cycleId
        - recipientId
        - amount
        - status
        - createdAt
    PayoutListResponse:
      type: object
      properties:
        payouts:
          type: array
          items:
            $ref: "#/components/schemas/Payout"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - payouts
        - total
        - page
        - limit
...